requirements.txt
alembic==1.12.0
aniso8601==9.0.1
annotated-types==0.6.0
attrs==23.1.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cors==1.0.1
coverage==7.3.2
cryptography==41.0.4
exceptiongroup==1.1.3
Faker==19.10.0
filelock==3.12.4
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Injector==0.15.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
flask-restplus==0.13.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
Flask-Testing==0.8.1
future==0.18.3
gevent==23.9.1
greenlet==3.0.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
injector==0.21.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
luhn==0.2.0
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
mirakuru==2.5.1
packaging==23.2
pkgutil_resolve_name==1.3.10
pluggy==1.3.0
port-for==0.7.2
psutil==5.9.6
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2==2.9.9
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
PySocks==1.7.1
pytest==7.4.2
pytest-cov==4.1.0
pytest-postgresql==5.0.0
python-creditcard @ git+https://github.com/maistodos/python-creditcard.git@6b6bcc5d975985337c35e5c1a4f87d861bcac731
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
requests-file==1.5.1
rpds-py==0.10.6
six==1.16.0
SQLAlchemy==2.0.21
tldextract==5.0.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.6
webargs==8.3.0
Werkzeug==3.0.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
